+++ app/src/main/java/com/example/android/birdsdaycounter/presentation/AddBirdDialog.kt
@@ -1,6 +1,5 @@
 package com.example.android.birdsdaycounter.presentation
 
-import android.app.Activity
 import android.app.Activity.RESULT_OK
 import android.content.ContentValues
 import android.content.Intent
@@ -14,9 +13,7 @@ import android.view.ViewGroup
 import android.view.WindowManager
 import android.widget.LinearLayout
 import android.widget.RadioButton
-import android.widget.Toast
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.content.ContentProviderCompat.requireContext
 import androidx.core.graphics.drawable.toBitmap
 import androidx.fragment.app.DialogFragment
 import androidx.lifecycle.viewModelScope
@@ -25,15 +22,15 @@ import com.example.android.birdsdaycounter.data.models.Bird
 import com.example.android.birdsdaycounter.databinding.AddBirdDialogBinding
 import com.example.android.birdsdaycounter.presentation.allBirdsFragment.AllBirdsViewModel
 import com.google.android.material.bottomsheet.BottomSheetDialog
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.cancel
 import kotlinx.coroutines.launch
 import java.io.ByteArrayOutputStream
-import java.io.File
-import java.io.FileOutputStream
 
 
 class AddBirdDialog(private val viewModel: AllBirdsViewModel) : DialogFragment() {
 
-    var cam_uri: Uri? = null
+    private var camUri: Uri? = null
     private lateinit var binding: AddBirdDialogBinding
     private var uri: Uri? = null
 
@@ -67,52 +64,32 @@ class AddBirdDialog(private val viewModel: AllBirdsViewModel) : DialogFragment()
         }
 
         binding.saveButton.setOnClickListener {
-             viewModel.viewModelScope.launch {
-                 val name = binding.birdNameET.text.toString()
-                 val age = binding.birdAgeET.text.toString()
-                 val id = binding.myRadioGroup.checkedRadioButtonId
-                 val gender = view.findViewById<RadioButton>(id).text.toString()
 
-                 val imgBitmap = binding.birdCreatImg.drawable.toBitmap()
-                 val bytes = ByteArrayOutputStream()
-                 imgBitmap.compress(Bitmap.CompressFormat.JPEG, 60, bytes)
+            val name = binding.birdNameET.text.toString()
+            val age = binding.birdAgeET.text.toString()
+            val id = binding.myRadioGroup.checkedRadioButtonId
+            val gender = view.findViewById<RadioButton>(id).text.toString()
+            val imgBitmap = binding.birdCreatImg.drawable.toBitmap()
 
-                 val bird = Bird(age, name, gender, null)
+                GlobalScope.launch {
+                    val bytes = ByteArrayOutputStream()
+                    imgBitmap.compress(Bitmap.CompressFormat.JPEG, 60, bytes)
 
-                 saveBird(bird, bytes)
-             }
+                    val bird = Bird(age, name, gender, null)
+
+                    saveBird(bird, bytes)
+                    cancel()
+                }
             this.dismiss()
-
         }
     }
 
     private fun saveBird(bird: Bird, byte: ByteArrayOutputStream) {
 
         viewModel.viewModelScope.launch {
-            saveBirdImage(bird, byte)
+            viewModel.saveBirdImage(bird, byte)
             viewModel.insertDB(bird)
             viewModel.newBirdWasAdded.value = true
-        }
-
-    }
-
-    private fun saveBirdImage(bird: Bird, byte: ByteArrayOutputStream) {
-        val imgLocation = requireActivity().filesDir.absolutePath + File.separator
-        val myAppDir = File(imgLocation)
-        if (!myAppDir.exists()) myAppDir.mkdir()
-
-        //creating child file
-        val fileName = "${System.currentTimeMillis()}.png"
-        val imageFile = File(myAppDir, fileName)
-        if (!imageFile.exists()) imageFile.createNewFile()
-
-        try {
-            val fo = FileOutputStream(imageFile)
-            fo.write(byte.toByteArray())
-            fo.close()
-
-            bird.imgLocation = imageFile.absolutePath
-        } catch (_: Exception) {
         }
 
     }
@@ -139,12 +116,12 @@ class AddBirdDialog(private val viewModel: AllBirdsViewModel) : DialogFragment()
         val values = ContentValues()
         values.put(MediaStore.Images.Media.TITLE, "New Picture")
         values.put(MediaStore.Images.Media.DESCRIPTION, "From Camera")
-        cam_uri = requireContext().contentResolver.insert(
+        camUri = requireContext().contentResolver.insert(
             MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
             values
         )
         val cameraIntent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
-        cameraIntent.putExtra(MediaStore.EXTRA_OUTPUT, cam_uri)
+        cameraIntent.putExtra(MediaStore.EXTRA_OUTPUT, camUri)
 
         cameraResultLauncher.launch(cameraIntent)
     }
@@ -157,10 +134,9 @@ class AddBirdDialog(private val viewModel: AllBirdsViewModel) : DialogFragment()
         galleryResultLauncher.launch(i)
     }
 
-
     private val galleryResultLauncher =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
-            if (result.resultCode == Activity.RESULT_OK) {
+            if (result.resultCode == RESULT_OK) {
                 val data = result.data
                 uri = data!!.data
                 binding.birdCreatImg.setImageURI(uri)
@@ -171,11 +147,10 @@ class AddBirdDialog(private val viewModel: AllBirdsViewModel) : DialogFragment()
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
 
             if (result.resultCode == RESULT_OK) {
-                if (cam_uri != null)
-                    binding.birdCreatImg.setImageURI(cam_uri)
+                if (camUri != null)
+                    binding.birdCreatImg.setImageURI(camUri)
             }
         }
-
 
     override fun onResume() {
         super.onResume()
